chore: moved usePublisher types to own file
